more optimization of expm_multiply_parallel

+++ quspin/tools/expm_multiply_parallel_core/expm_multiply_parallel_core.py
@@ -162,7 +162,7 @@ class expm_multiply_parallel(object):
             if work_array.dtype != v_dtype:
                 raise ValueError("work_array must be array of dtype which matches the result of the matrix-vector multiplication.")
 
-        a = _np.array(self._a,dtype=a_dtype)
+        a = _np.array(self._a,dtype=v_dtype)
         _wrapper_expm_multiply(self._A.indptr,self._A.indices,self._A.data,
                     self._m_star,self._s,a,self._tol,self._mu,v,work_array)
